+++ src/nylas.ts
@@ -8,6 +8,11 @@ import { Applications } from './resources/applications.js';
 import { Messages } from './resources/messages.js';
 import { Drafts } from './resources/drafts.js';
 import { Threads } from './resources/threads.js';
+import { Connectors } from './resources/connectors.js';
+import { Folders } from './resources/folders.js';
+import { Grants } from './resources/grants.js';
+import { Contacts } from './resources/contacts.js';
+import { Attachments } from './resources/attachments.js';
 
 /**
  * The entry point to the Node SDK
@@ -21,6 +26,10 @@ export default class Nylas {
    */
   public applications: Applications;
   /**
+   * Access the Attachments API
+   */
+  public attachments: Attachments;
+  /**
    * Access the Auth API
    */
   public auth: Auth;
@@ -29,6 +38,14 @@ export default class Nylas {
    */
   public calendars: Calendars;
   /**
+   * Access the Connectors API
+   */
+  public connectors: Connectors;
+  /**
+   * Access the Contacts API
+   */
+  public contacts: Contacts;
+  /**
    * Access the Drafts API
    */
   public drafts: Drafts;
@@ -36,6 +53,10 @@ export default class Nylas {
    * Access the Events API
    */
   public events: Events;
+  /**
+   * Access the Grants API
+   */
+  public grants: Grants;
   /**
    * Access the Messages API
    */
@@ -48,6 +69,10 @@ export default class Nylas {
    * Access the Webhooks API
    */
   public webhooks: Webhooks;
+  /**
+   * Access the Folders API
+   */
+  public folders: Folders;
 
   /**
    * The configured API client
@@ -68,11 +93,16 @@ export default class Nylas {
     this.applications = new Applications(this.apiClient);
     this.auth = new Auth(this.apiClient);
     this.calendars = new Calendars(this.apiClient);
+    this.connectors = new Connectors(this.apiClient);
     this.drafts = new Drafts(this.apiClient);
     this.events = new Events(this.apiClient);
+    this.grants = new Grants(this.apiClient);
     this.messages = new Messages(this.apiClient);
     this.threads = new Threads(this.apiClient);
     this.webhooks = new Webhooks(this.apiClient);
+    this.folders = new Folders(this.apiClient);
+    this.contacts = new Contacts(this.apiClient);
+    this.attachments = new Attachments(this.apiClient);
 
     return this;
   }