+++ src/resources/resource.ts
@@ -192,6 +192,32 @@ export class Resource {
       overrides,
     });
   }
+
+  protected _getRaw({
+    path,
+    queryParams,
+    overrides,
+  }: FindParams<void>): Promise<Buffer> {
+    return this.apiClient.requestRaw({
+      method: 'GET',
+      path,
+      queryParams,
+      overrides,
+    });
+  }
+
+  protected _getStream({
+    path,
+    queryParams,
+    overrides,
+  }: FindParams<void>): Promise<NodeJS.ReadableStream> {
+    return this.apiClient.requestStream({
+      method: 'GET',
+      path,
+      queryParams,
+      overrides,
+    });
+  }
 }
 
 type ListYieldReturn<T> = T & {