+++ tests/nylas.spec.ts
@@ -20,6 +20,8 @@ describe('Nylas', () => {
       expect(nylas.calendars.constructor.name).toBe('Calendars');
       expect(nylas.events.constructor.name).toBe('Events');
       expect(nylas.webhooks.constructor.name).toBe('Webhooks');
+      expect(nylas.folders.constructor.name).toBe('Folders');
+      expect(nylas.attachments.constructor.name).toBe('Attachments');
     });
 
     it('should configure the apiClient', () => {