+++ tests/resources/auth.spec.ts
@@ -45,6 +45,27 @@ describe('Auth', () => {
         });
       });
 
+      it('should default clientSecret to the API key', async () => {
+        const payload: CodeExchangeRequest = {
+          clientId: 'clientId',
+          redirectUri: 'https://redirect.uri/path',
+          code: 'code',
+        };
+        await auth.exchangeCodeForToken(payload);
+
+        expect(apiClient.request).toHaveBeenCalledWith({
+          method: 'POST',
+          path: '/v3/connect/token',
+          body: {
+            clientId: 'clientId',
+            clientSecret: 'apiKey',
+            redirectUri: 'https://redirect.uri/path',
+            code: 'code',
+            grantType: 'authorization_code',
+          },
+        });
+      });
+
       it('should set codeVerifier', async () => {
         const payload: CodeExchangeRequest = {
           clientId: 'clientId',
@@ -92,37 +113,62 @@ describe('Auth', () => {
           },
         });
       });
+
+      it('should default clientSecret to the API key', async () => {
+        const payload: TokenExchangeRequest = {
+          clientId: 'clientId',
+          redirectUri: 'https://redirect.uri/path',
+          refreshToken: 'refreshToken',
+        };
+        await auth.refreshAccessToken(payload);
+
+        expect(apiClient.request).toHaveBeenCalledWith({
+          method: 'POST',
+          path: '/v3/connect/token',
+          body: {
+            clientId: 'clientId',
+            clientSecret: 'apiKey',
+            redirectUri: 'https://redirect.uri/path',
+            refreshToken: 'refreshToken',
+            grantType: 'refresh_token',
+          },
+        });
+      });
     });
   });
-  // describe('Validating token', () => {
-  //   describe('validateIDToken', () => {
-  //     it('should call apiClient.request with the correct params', async () => {
-  //       await auth.validateIDToken('id123');
-  //
-  //       expect(apiClient.request).toHaveBeenCalledWith({
-  //         method: 'GET',
-  //         path: '/v3/connect/tokeninfo',
-  //         queryParams: {
-  //           idToken: 'id123',
-  //         },
-  //       });
-  //     });
-  //   });
-  //
-  //   describe('validateAccessToken', () => {
-  //     it('should call apiClient.request with the correct params', async () => {
-  //       await auth.validateAccessToken('accessToken123');
-  //
-  //       expect(apiClient.request).toHaveBeenCalledWith({
-  //         method: 'GET',
-  //         path: '/v3/connect/tokeninfo',
-  //         queryParams: {
-  //           accessToken: 'accessToken123',
-  //         },
-  //       });
-  //     });
-  //   });
-  // });
+  describe('customAuthentication', () => {
+    it('should call apiClient.request with the correct params', async () => {
+      await auth.customAuthentication({
+        requestBody: {
+          provider: 'google',
+          settings: {
+            test_setting: 'abc123',
+          },
+          scope: ['calendar'],
+          state: 'state123',
+        },
+        overrides: {
+          apiUri: 'https://test.api.nylas.com',
+        },
+      });
+
+      expect(apiClient.request).toHaveBeenCalledWith({
+        method: 'POST',
+        path: '/v3/connect/custom',
+        body: {
+          provider: 'google',
+          settings: {
+            test_setting: 'abc123',
+          },
+          scope: ['calendar'],
+          state: 'state123',
+        },
+        overrides: {
+          apiUri: 'https://test.api.nylas.com',
+        },
+      });
+    });
+  });
   describe('URL building', () => {
     describe('urlForAuthentication', () => {
       it('should build the correct url', () => {